refactor(tui): split view.rs into responsibility submodules
view.rs grew to 995 lines and the file-size-guard hook started
blocking edits to it. Split by responsibility, no behaviour
change:

- view.rs                 orchestrator: `render_app`, `render_main`,
                          overlay dispatch. ~200 lines.
- view/inline.rs          span-level markdown (pretty + raw
                          highlight) and TODO prefix split.
- view/outline.rs         outline-tree rendering — `render_outline`,
                          `render_block`, `emit_block_lines`,
                          `RenderMode`, `CursorStyle`.
- view/overlays.rs        every modal popup: quick switcher,
                          search, slash, command bar, error,
                          help, inline autocomplete.

Each file lands well under the 600-line warning threshold,
leaving room for the inline backlinks section added next.
