+++ crates/outl-tui/src/view/outline.rs
@@ -1,0 +1,283 @@
+//! Outline rendering — turn a `ParsedPage` (current view) into a
+//! flat `Vec<Line>` for ratatui, with selection / cursor / TODO
+//! decoration.
+
+use crate::outline_ops::path_for_index;
+use crate::state::{App, Focus, Mode};
+use crate::theme::Theme;
+use crate::view::inline::{highlight_inline, render_markdown_inline, split_todo_prefix};
+use outl_md::inline::byte_index_for_char;
+use outl_md::parse::{OutlineNode, ParsedPage};
+use outl_md::view::{block_to_rows, BlockRowKind};
+use ratatui::style::Style;
+use ratatui::text::{Line, Span};
+
+/// Render the outline into a flat list of `Line`s for ratatui, and
+/// report the visual line index where the *selected* block's bullet
+/// row landed. The caller uses that index to keep the selection
+/// inside the scrolled viewport.
+pub(crate) fn render_outline(p: &ParsedPage, app: &App) -> (Vec<Line<'static>>, Option<usize>) {
+    let mut out = Vec::new();
+    for (k, v) in &p.properties {
+        out.push(Line::from(vec![
+            Span::styled(format!("{k}:: "), app.theme.property_key),
+            Span::styled(v.clone(), app.theme.property_value),
+        ]));
+    }
+    if !p.properties.is_empty() && !p.blocks.is_empty() {
+        out.push(Line::from(""));
+    }
+    let mut cursor = 0usize;
+    let mut selected_line: Option<usize> = None;
+    for block in &p.blocks {
+        render_block(block, 0, &mut cursor, app, &mut out, &mut selected_line);
+    }
+    (out, selected_line)
+}
+
+pub(crate) fn render_block(
+    b: &OutlineNode,
+    indent: u32,
+    cursor: &mut usize,
+    app: &App,
+    out: &mut Vec<Line<'static>>,
+    selected_line: &mut Option<usize>,
+) {
+    // Outline only owns selection/cursor decoration when focus lives
+    // here. With `Focus::Backlink`, the bullet/caret belong to the
+    // backlinks section — drawing them on the outline too would leave
+    // a "ghost cursor" on the last outline block (the value `selected`
+    // still happens to point at).
+    let focused_on_outline = matches!(app.focus, Focus::Outline);
+    let is_selected = focused_on_outline && *cursor == app.selected;
+    // Record the visual line where this block's bullet row begins so
+    // the caller can scroll the viewport to keep it visible.
+    if is_selected && selected_line.is_none() {
+        *selected_line = Some(out.len());
+    }
+    let in_visual_range = focused_on_outline
+        && app
+            .visual_range()
+            .is_some_and(|(lo, hi)| *cursor >= lo && *cursor <= hi);
+    let editing_here = focused_on_outline
+        && matches!(&app.mode, Mode::Insert { block_path, .. }
+            if path_for_index(&app.page.blocks, *cursor).as_deref() == Some(block_path.as_slice()));
+
+    let bullet_style = if is_selected || in_visual_range {
+        app.theme.selected_bullet
+    } else {
+        app.theme.bullet
+    };
+
+    // Determine which text and cursor position to render. Three cases:
+    //   1. Editing here       → buffer with caret cursor.
+    //   2. Selected in Normal → block text with block-style cursor.
+    //   3. Anything else      → block text, no cursor, pretty render.
+    let mode = if editing_here {
+        if let Mode::Insert { buffer, .. } = &app.mode {
+            RenderMode::Editing {
+                text: buffer.as_string(),
+                cursor_char: buffer.cursor,
+            }
+        } else {
+            unreachable!("editing_here matched but mode isn't Insert")
+        }
+    } else if is_selected && matches!(app.mode, Mode::Normal) {
+        RenderMode::NormalCursor {
+            text: b.text.clone(),
+            cursor_char: app.cursor_col,
+        }
+    } else {
+        RenderMode::Pretty {
+            text: b.text.clone(),
+        }
+    };
+
+    let has_auto_run = b.properties.iter().any(|(k, _)| k == "auto-run");
+    emit_block_lines(indent, bullet_style, &mode, has_auto_run, app, out);
+
+    for (k, v) in &b.properties {
+        let mut prop_spans: Vec<Span<'_>> = Vec::new();
+        for _ in 0..indent {
+            prop_spans.push(Span::styled("│ ", app.theme.dim));
+        }
+        prop_spans.push(Span::raw("  ".to_string()));
+        prop_spans.push(Span::styled(format!("{k}:: "), app.theme.property_key));
+        prop_spans.push(Span::styled(v.clone(), app.theme.property_value));
+        out.push(Line::from(prop_spans));
+    }
+    *cursor += 1;
+    for child in &b.children {
+        render_block(child, indent + 1, cursor, app, out, selected_line);
+    }
+}
+
+/// Where the cursor sits on a block being rendered, and what style
+/// the renderer should use for it. The UI-agnostic decomposition
+/// lives in [`outl_md::view`]; this enum carries the *TUI-flavored*
+/// detail of "caret vs block cursor".
+pub(crate) enum RenderMode {
+    /// Insert mode — show the live buffer with a thin caret at
+    /// `cursor_char`. Markdown is rendered raw so columns match bytes.
+    Editing { text: String, cursor_char: usize },
+    /// Normal mode on the selected block — show a vim-style block
+    /// cursor on the character under `cursor_char`. Raw render.
+    NormalCursor { text: String, cursor_char: usize },
+    /// Anything else — markdown is rendered prettily; no cursor.
+    Pretty { text: String },
+}
+
+/// Emit one or more ratatui [`Line`]s for a block's text.
+///
+/// Decomposition into visual rows (bullet vs continuation vs code
+/// fence marker vs code fence body) is delegated to
+/// [`outl_md::view::block_to_rows`] so the Tauri GUI and mobile
+/// clients use the same classification. This function is the
+/// TUI-specific mapping: each [`outl_md::view::BlockRow`] becomes a
+/// `Line` of `Span`s using the active theme.
+pub(crate) fn emit_block_lines(
+    indent: u32,
+    bullet_style: Style,
+    mode: &RenderMode,
+    has_auto_run: bool,
+    app: &App,
+    out: &mut Vec<Line<'static>>,
+) {
+    let (text, cursor_char, cursor_style) = match mode {
+        RenderMode::Editing { text, cursor_char } => {
+            (text.as_str(), Some(*cursor_char), Some(CursorStyle::Caret))
+        }
+        RenderMode::NormalCursor { text, cursor_char } => {
+            (text.as_str(), Some(*cursor_char), Some(CursorStyle::Block))
+        }
+        RenderMode::Pretty { text } => (text.as_str(), None, None),
+    };
+    let pretty = matches!(mode, RenderMode::Pretty { .. });
+    let rows = block_to_rows(text, indent, cursor_char);
+
+    // TODO/DONE checkbox decoration only fits on single-line bullets
+    // (multi-line ones would have the icon floating above body text).
+    let single_line_pretty = pretty && rows.len() == 1;
+
+    for row in &rows {
+        let mut spans: Vec<Span<'_>> = Vec::new();
+        for _ in 0..row.indent {
+            spans.push(Span::styled("│ ", app.theme.dim));
+        }
+        match row.kind {
+            BlockRowKind::Bullet => {
+                // Blocks with `auto-run::` get a ⚡ before the bullet
+                // so the user can see at a glance which cells re-run
+                // themselves on page open.
+                if has_auto_run {
+                    spans.push(Span::styled("⚡", app.theme.hint));
+                }
+                spans.push(Span::styled("- ", bullet_style));
+            }
+            BlockRowKind::Continuation
+            | BlockRowKind::CodeFenceMarker
+            | BlockRowKind::CodeFenceBody => {
+                // Indent the continuation rows by the same width
+                // the ⚡ added on the bullet row so columns align.
+                if has_auto_run {
+                    spans.push(Span::raw(" "));
+                }
+                spans.push(Span::raw("  "));
+            }
+        }
+
+        // If the cursor is on this row we always go raw — we want
+        // bytes to line up with what the user typed, regardless of
+        // fence state.
+        if let (Some(col), Some(style)) = (row.cursor_col, cursor_style) {
+            emit_row_with_cursor(row.text, col, style, &app.theme, &mut spans);
+        } else {
+            // A bullet row whose text opens a code fence (`` ```lisp ``)
+            // is *both* a bullet and a fence marker — style the text
+            // dimly so the fence reads visually like the rest of the
+            // code block while keeping the `- ` glyph emitted above.
+            let bullet_is_fence_opener = matches!(row.kind, BlockRowKind::Bullet)
+                && row.text.trim_start().starts_with("```");
+            match row.kind {
+                _ if pretty && bullet_is_fence_opener => {
+                    spans.push(Span::styled(row.text.to_string(), app.theme.dim));
+                }
+                BlockRowKind::CodeFenceMarker if pretty => {
+                    spans.push(Span::styled(row.text.to_string(), app.theme.dim));
+                }
+                BlockRowKind::CodeFenceBody if pretty => {
+                    spans.push(Span::styled(row.text.to_string(), app.theme.code));
+                }
+                BlockRowKind::Bullet if single_line_pretty => {
+                    let (todo_state, body) = split_todo_prefix(row.text);
+                    match todo_state {
+                        Some(false) => {
+                            spans.push(Span::styled("☐ ", app.theme.todo_open));
+                            spans.extend(render_markdown_inline(body, &app.theme, &app.index));
+                        }
+                        Some(true) => {
+                            spans.push(Span::styled("☑ ", app.theme.todo_done));
+                            for sp in render_markdown_inline(body, &app.theme, &app.index) {
+                                spans.push(Span::styled(
+                                    sp.content.into_owned(),
+                                    sp.style.patch(app.theme.todo_done_body),
+                                ));
+                            }
+                        }
+                        None => {
+                            spans.extend(render_markdown_inline(row.text, &app.theme, &app.index))
+                        }
+                    }
+                }
+                _ => spans.extend(render_markdown_inline(row.text, &app.theme, &app.index)),
+            }
+        }
+        out.push(Line::from(spans));
+    }
+}
+
+/// Draw one row with the cursor highlighted at `col` (a char index
+/// into `text`). Splits the row in three: left of cursor, the char
+/// under the cursor (or a thin caret if past-end), right of cursor.
+fn emit_row_with_cursor(
+    text: &str,
+    col: usize,
+    style: CursorStyle,
+    theme: &Theme,
+    spans: &mut Vec<Span<'static>>,
+) {
+    let byte = byte_index_for_char(text, col);
+    let (left, right) = text.split_at(byte);
+    spans.extend(highlight_inline(left, theme));
+    let mut right_chars = right.chars();
+    match (right_chars.next(), style) {
+        (Some(ch), CursorStyle::Caret) => {
+            // Thin caret BEFORE the next char.
+            spans.push(Span::styled("▏", theme.cursor_caret));
+            spans.push(Span::raw(ch.to_string()));
+            let rest: String = right_chars.collect();
+            spans.extend(highlight_inline(&rest, theme));
+        }
+        (Some(ch), CursorStyle::Block) => {
+            // Inverted-color block cursor on the char under it.
+            spans.push(Span::styled(ch.to_string(), theme.cursor_block));
+            let rest: String = right_chars.collect();
+            spans.extend(highlight_inline(&rest, theme));
+        }
+        (None, CursorStyle::Caret) => {
+            spans.push(Span::styled("▏", theme.cursor_caret));
+        }
+        (None, CursorStyle::Block) => {
+            spans.push(Span::styled("▏", theme.cursor_block));
+        }
+    }
+}
+
+/// Cursor visual style. `Caret` is the thin `▏` (Insert mode);
+/// `Block` is the inverted single-char box (Normal mode on the
+/// selected block).
+#[derive(Debug, Clone, Copy)]
+enum CursorStyle {
+    Caret,
+    Block,
+}
